agents_v2/pyproject.toml
[project]
name = "agents-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "azure-ai-projects==2.0.0b2",
    "azure-identity>=1.25.0",
    "ipykernel>=6.30.1",
    "jsonref>=1.1.0",
    "openai>=2.9.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]